+++ txcdr-server/src/graphql/auth.ts
@@ -1,6 +1,6 @@
 import { $Enums, PrismaClient, User } from '@prisma/client';
 import { AuthUser, createClient } from '@supabase/supabase-js';
-import { supabaseKey, supabaseUrl } from './utils/config.js';
+import { supabaseKey, supabaseUrl } from './config.js';
 
 const prisma = new PrismaClient();
 const supabase = createClient(supabaseUrl, supabaseKey);
